requirements.txt
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pyzotero>=1.6.11
html2text==2020.1.16
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.28.0
pathlib2>=2.3.0
pydub>=0.25.1 